refactor: preprocessor obtains array of bad windows
The module iterates through all LogNodes that are labelled with 'F' to compile a sequence of labels within a 3 hour window. If the last label of the window is 'B' the window is labelled as a bad window, otherwise it is labelled as a good window (subject to change).

Daniel